Moved death_functions.py to helper_functions package
